rename rpc module to rust_rpc to avoid naming conflicts
