+++ Makefile.toml
@@ -213,7 +213,8 @@ env = { "RUST_BACKTRACE" = "1", "WASMTIME_BACKTRACE_DETAILS" = "1", "RUST_LOG" =
 script = '''
 cargo test --package golem-worker-executor-base --test integration scalability -- --nocapture
 cargo test --package golem-worker-executor-base --test integration hot_update -- --nocapture
-cargo test --package golem-worker-executor-base --test integration rpc -- --nocapture
+cargo test --package golem-worker-executor-base --test integration rust_rpc -- --nocapture
+
 '''
 
 [tasks.worker-executor-tests-group5]
