+++ golem-worker-executor-base/tests/rust_rpc.rs
